+++ admin/components.php
@@ -123,6 +123,22 @@ if (count($componentsec) != 0) {
 		$submitclass = 'hidden';
 		
 	}
+# register and queue CodeMirror files
+if (!getDef('GSNOHIGHLIGHT',true)){
+  register_script('codemirror', $SITEURL.$GSADMIN.'/template/js/codemirror/lib/codemirror-compressed.js', '0.2.0', FALSE);
+  register_style('codemirror-css',$SITEURL.$GSADMIN.'/template/js/codemirror/lib/codemirror.css','screen',FALSE);
+  register_style('codemirror-theme',$SITEURL.$GSADMIN.'/template/js/codemirror/theme/default.css','screen',FALSE);
+
+  queue_script('codemirror', GSBACK);
+  queue_style('codemirror-css', GSBACK);
+  queue_style('codemirror-theme', GSBACK);
+
+  function set_gs_codemirror() {
+    echo "<script>GS.CodeMirror = new Array(); GS.CodeMirror['enabled'] = true; GS.CodeMirror['mode'] = 'application/x-httpd-php';</script>".PHP_EOL;
+  }
+
+  add_action('header', 'set_gs_codemirror');
+}
 
 get_template('header', cl($SITENAME).' &raquo; '.i18n_r('COMPONENTS')); 
 
@@ -144,7 +160,7 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('COMPONENTS'));
 		<input type="hidden" id="id" value="<?php echo $count; ?>" />
 		<input type="hidden" id="nonce" name="nonce" value="<?php echo get_nonce("modify_components"); ?>" />
 
-		<div id="divTxt"></div> 
+		<div id="divTxt"></div>
 		<?php echo $table; ?>
 		<p id="submit_line" class="<?php echo $submitclass; ?>" >
 			<span><input type="submit" class="submit" name="submitted" id="button" value="<?php i18n('SAVE_COMPONENTS');?>" /></span> &nbsp;&nbsp;<?php i18n('OR'); ?>&nbsp;&nbsp; <a class="cancel" href="components.php?cancel"><?php i18n('CANCEL'); ?></a>