+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * Added `SITEDESC` website field to store website description
     * Added `get_site_description()` theme function to get value from `SITEDESC` website field
     * Added `DESC` user field to store user description (e.g. user biography) for public display
+    * Added code highlight to components
 
 1. [](#improved)
     * Updated theme function `get_navigation()` to add `aria-role="page"` to current page in menu